+++ packages/manager/src/features/Linodes/LinodesCreate/LinodeCreate.tsx
@@ -68,13 +68,13 @@ import {
   StackScriptFormStateHandlers,
   TypeInfo,
   WithDisplayData,
-  WithLinodesProps,
   WithTypesRegionsAndImages,
 } from './types';
 import { UserDataAccordion } from './UserDataAccordion/UserDataAccordion';
 import { extendType } from 'src/utilities/extendType';
 import { WithTypesProps } from 'src/containers/types.container';
 import { RegionsProps } from 'src/containers/regions.container';
+import { WithLinodesProps } from 'src/containers/withLinodes.container';
 
 type ClassNames =
   | 'form'
